requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
dj-database-url==2.1.0
Django==4.2.5
django-filter==23.3
django-heroku==0.3.1
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
packaging==23.1
phonenumbers==8.13.20
Pillow==10.0.1
psycopg2==2.9.7
psycopg2-binary==2.9.7
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.5
whitenoise==6.5.0